Change "full" to "base" and remove unused code
